pthread/samples: Fix errors on alpinelinux
pthread_barrier_init.c: In function 'test_task_fn':
pthread_barrier_init.c:12:25: error: format '%ld' expects argument of type 'long int', but argument 2 has type 'pthread_t' {aka 'struct __pthread *'} [-Werror=format=]
  printf("test_task_fn %ld.\n", pthread_self());
                       ~~^      ~~~~~~~~~~~~~~
pthread_barrier_init.c:14:23: error: invalid operands to binary % (have 'pthread_t' {aka 'struct __pthread *'} and 'int')
  sleep(pthread_self() % NR_THREAD);
        ~~~~~~~~~~~~~~ ^
pthread_barrier_init.c:18:19: error: format '%ld' expects argument of type 'long int', but argument 2 has type 'pthread_t' {aka 'struct __pthread *'} [-Werror=format=]
  printf("thread %ld done.\n", pthread_self());
                 ~~^           ~~~~~~~~~~~~~~
cc1: all warnings being treated as errors
make: *** [<builtin>: pthread_barrier_init.o] Error 1

+++ glibc/pthread/samples/pthread_barrier_init.c
@@ -9,13 +9,13 @@ pthread_barrierattr_t barrierattr;
 
 void* test_task_fn(void* unused)
 {
-	printf("test_task_fn %ld.\n", pthread_self());
+	printf("test_task_fn %ld.\n", (unsigned long)pthread_self());
 
-	sleep(pthread_self() % NR_THREAD);
+	sleep((unsigned long)pthread_self() % NR_THREAD);
 
 	pthread_barrier_wait(&barrier);
 
-	printf("thread %ld done.\n", pthread_self());
+	printf("thread %ld done.\n", (unsigned long)pthread_self());
 
 	pthread_exit(NULL);
 	return NULL;
